+++ R/RcppExports.R
@@ -25,15 +25,15 @@ weibullDensity <- function(x, beta, lambda, censored) {
     .Call(`_weibulltools_weibullDensity`, x, beta, lambda, censored)
 }
 
-LikelihoodWeibull <- function(x, parameter, status, prior) {
-    .Call(`_weibulltools_LikelihoodWeibull`, x, parameter, status, prior)
+LikelihoodWeibull <- function(x, parameter, status, prior, P, logL) {
+    invisible(.Call(`_weibulltools_LikelihoodWeibull`, x, parameter, status, prior, P, logL))
 }
 
 normalize <- function(M) {
     invisible(.Call(`_weibulltools_normalize`, M))
 }
 
-#' EM-Algorithm using Newton-Raphson method
+#' EM-Algorithm using Newton-Raphson Method
 #'
 #' This method uses the EM-Algorithm to estimate the parameters of a univariate
 #' mixture model. Until now, the mixture model can consist of k two-parametric
@@ -55,12 +55,13 @@ normalize <- function(M) {
 #' @param post a numeric matrix specifiying initial a-posteriori probabilities.
 #'   The number of rows have to be in line with observations \code{x} and the
 #'   number of columns must equal the mixture components \code{k}.
-#' @param distribution supposed mixture model. If \code{"weibull"} is used. Other
-#'   distributions have not been implemented yet.
+#' @param distribution supposed distribution of mixture model components.
+#'   The value must be \code{"weibull"}. Other distributions have not been
+#'   implemented yet.
 #' @param k integer of mixture components, default is 2.
-#' @param method default method is \code{"EM"}. Other methods have not been implemented
-#'   yet.
-#' @param n_iter numeric value defining the maximum number of iterations.
+#' @param method default method is \code{"EM"}. Other methods have not been
+#'   implemented yet.
+#' @param n_iter integer defining the maximum number of iterations.
 #' @param conv_limit numeric value defining the convergence limit.
 #'
 #' @return Returns a list with the following components:
@@ -69,8 +70,9 @@ normalize <- function(M) {
 #'     first row the estimated scale parameters \eqn{\eta} and in the second the
 #'     estimated shape parameters \eqn{\beta} are provided. The first column belongs
 #'     to the first mixture component and so forth.
-#'   \item \code{posteriori} : A matrix with a-posteriori probabilities.
-#'   \item \code{priori} : A vector with a-priori probabilities.}
+#'   \item \code{posteriori} : A matrix with estimated a-posteriori probabilities.
+#'   \item \code{priori} : A vector with estimated a-priori probabilities.
+#'   \item \code{logL} : The value of the complete log-likelihood.}
 #' @export
 #' @examples
 #' # Data is taken from given reference:
@@ -85,13 +87,13 @@ normalize <- function(M) {
 #'          0, 1, 1, 1, 1, 1, 1)
 #' post_dirichlet <- LearnBayes::rdirichlet(n = length(hours),
 #'                                          par = rep(.1, 2))
-#' mix_mod_em_cpp <- mixture_em_cpp(x = hours,
-#'                                  event = state,
-#'                                  post = post_dirichlet,
-#'                                  distribution = "weibull",
-#'                                  k = 2,
-#'                                  method = "EM",
-#'                                  n_iter = 150)
+#' mix_mod_em <- mixture_em_cpp(x = hours,
+#'                              event = state,
+#'                              post = post_dirichlet,
+#'                              distribution = "weibull",
+#'                              k = 2,
+#'                              method = "EM",
+#'                              n_iter = 150)
 #'
 mixture_em_cpp <- function(x, event, post, distribution = "weibull", k = 2L, method = "EM", n_iter = 100L, conv_limit = 1e-6) {
     .Call(`_weibulltools_mixture_em_cpp`, x, event, post, distribution, k, method, n_iter, conv_limit)