+++ src/em_mixture.cpp
@@ -9,7 +9,7 @@ double getLambda(NumericVector& x, NumericVector& weights, NumericVector& status
                  double beta) {
 
   NumericVector xBeta = pow(x, beta) * weights;
-  NumericVector failuresWeights = weights[status ==1];
+  NumericVector failuresWeights = weights[status == 1];
 
   double lambda = sum(xBeta) / sum(failuresWeights);
   lambda = std::pow(lambda, 1.0 / beta);
@@ -28,7 +28,7 @@ double g(NumericVector& x,NumericVector& weights, NumericVector& status,
   logFailures = logFailures * failuresWeights;
 
   NumericVector l = log(x);
-  NumericVector xBeta = pow(x, beta);
+  NumericVector xBeta =pow(x, beta);
   xBeta = xBeta * weights;
 
   NumericVector xLog = xBeta * l;
@@ -38,9 +38,10 @@ double g(NumericVector& x,NumericVector& weights, NumericVector& status,
 // [[Rcpp::export]]
 double gDiv(NumericVector& x, NumericVector& weights, double beta) {
 
+  // double n = x.length();
   NumericVector l = log(x);
   NumericVector l2 = pow(l, 2);
-  NumericVector xBeta = pow(x, beta);
+  NumericVector xBeta =pow(x, beta);
   xBeta = xBeta * weights;
 
   NumericVector xBetaLog = xBeta * l;
@@ -58,16 +59,14 @@ double NewtonRaphson(NumericVector& x, NumericVector& weights,
 
   double beta = 1;
   double alpha = 1;
-  for(int i = 0; i < 10; i++) {
+  for (int i = 0; i < 10; i++) {
     beta = exp(alpha);
-    alpha = alpha- g(x, weights, status, beta) / (gDiv(x,weights,beta) * beta);
-    //beta = beta- g(x, weights, status, beta)/gDiv(x,weights,beta);
+    alpha = alpha- g(x, weights, status, beta) / (gDiv(x, weights, beta) * beta);
+    //beta = beta- g(x, weights, status, beta) / gDiv(x,weights,beta);
   }
   beta = exp(alpha);
   return beta;
 }
-
-
 
 // [[Rcpp::export]]
 NumericMatrix MStepWeibull(NumericVector& x, NumericMatrix& posterior,
@@ -82,12 +81,12 @@ NumericMatrix MStepWeibull(NumericVector& x, NumericMatrix& posterior,
     double beta_i = NewtonRaphson(x, p_i, status);
     double lambda_i = getLambda(x, p_i, status, beta_i);
 
+    // Test lambda und beta vertauscht
     parameter(0, i) = lambda_i;
     parameter(1, i) = beta_i;
   }
   return parameter;
 }
-
 
 // [[Rcpp::export]]
 double weibullDensity(double x, double beta, double lambda, double censored) {
@@ -96,7 +95,7 @@ double weibullDensity(double x, double beta, double lambda, double censored) {
   double s = pow(y, beta - 1.0);
   double t = pow(y, beta);
 
-  if (censored < 1.0) {
+  if(censored < 1.0) {
     return exp(-t);
   }
   return (beta / lambda) * s * exp(-t);
@@ -111,31 +110,32 @@ void LikelihoodWeibull(NumericVector& x, NumericMatrix& parameter,
 
   for (int i = 0; i < x.length(); i++) {
     for (int j = 0; j < nMixures; j++) {
-      double p = weibullDensity(x(i), parameter(0, j), parameter(1, j), status(i));
-      logL(i, j) = log(p);
-      P(i, j) = p * prior(j);
+      double p = weibullDensity(x(i), parameter(1, j), parameter(0, j), status(i));
+      logL(i,j) = log(p);
+      P(i,j) = p * prior(j);
     }
   }
 }
 
 // [[Rcpp::export]]
 void normalize(NumericMatrix& M) {
-  NumericVector norm = rowSums(M);
 
+  NumericVector norm = rowSums(M);
   for (int i = 0; i < M.rows(); ++i) {
     M(i,_) = M(i,_) / norm[i];
   }
 }
 
-// [[Rcpp::export]]
-double logLikelihood(arma::mat& posterior, arma::mat& logDensity, arma::vec& prior){
+double logLikelihood(arma::mat& posterior, arma::mat& logDensity,
+                     arma::vec& prior) {
+
   double Q = arma::accu(posterior % logDensity);
   double r = arma::accu(posterior * prior);
 
   return Q + r;
 }
 
-//' EM-Algorithm using Newton-Raphson method
+//' EM-Algorithm using Newton-Raphson Method
 //'
 //' This method uses the EM-Algorithm to estimate the parameters of a univariate
 //' mixture model. Until now, the mixture model can consist of k two-parametric
@@ -157,12 +157,13 @@ double logLikelihood(arma::mat& posterior, arma::mat& logDensity, arma::vec& pri
 //' @param post a numeric matrix specifiying initial a-posteriori probabilities.
 //'   The number of rows have to be in line with observations \code{x} and the
 //'   number of columns must equal the mixture components \code{k}.
-//' @param distribution supposed mixture model. If \code{"weibull"} is used. Other
-//'   distributions have not been implemented yet.
+//' @param distribution supposed distribution of mixture model components.
+//'   The value must be \code{"weibull"}. Other distributions have not been
+//'   implemented yet.
 //' @param k integer of mixture components, default is 2.
-//' @param method default method is \code{"EM"}. Other methods have not been implemented
-//'   yet.
-//' @param n_iter numeric value defining the maximum number of iterations.
+//' @param method default method is \code{"EM"}. Other methods have not been
+//'   implemented yet.
+//' @param n_iter integer defining the maximum number of iterations.
 //' @param conv_limit numeric value defining the convergence limit.
 //'
 //' @return Returns a list with the following components:
@@ -171,8 +172,9 @@ double logLikelihood(arma::mat& posterior, arma::mat& logDensity, arma::vec& pri
 //'     first row the estimated scale parameters \eqn{\eta} and in the second the
 //'     estimated shape parameters \eqn{\beta} are provided. The first column belongs
 //'     to the first mixture component and so forth.
-//'   \item \code{posteriori} : A matrix with a-posteriori probabilities.
-//'   \item \code{priori} : A vector with a-priori probabilities.}
+//'   \item \code{posteriori} : A matrix with estimated a-posteriori probabilities.
+//'   \item \code{priori} : A vector with estimated a-priori probabilities.
+//'   \item \code{logL} : The value of the complete log-likelihood.}
 //' @export
 //' @examples
 //' # Data is taken from given reference:
@@ -187,45 +189,38 @@ double logLikelihood(arma::mat& posterior, arma::mat& logDensity, arma::vec& pri
 //'          0, 1, 1, 1, 1, 1, 1)
 //' post_dirichlet <- LearnBayes::rdirichlet(n = length(hours),
 //'                                          par = rep(.1, 2))
-//' mix_mod_em_cpp <- mixture_em_cpp(x = hours,
-//'                                  event = state,
-//'                                  post = post_dirichlet,
-//'                                  distribution = "weibull",
-//'                                  k = 2,
-//'                                  method = "EM",
-//'                                  n_iter = 150)
+//' mix_mod_em <- mixture_em_cpp(x = hours,
+//'                              event = state,
+//'                              post = post_dirichlet,
+//'                              distribution = "weibull",
+//'                              k = 2,
+//'                              method = "EM",
+//'                              n_iter = 150)
 //'
 // [[Rcpp::export]]
 List mixture_em_cpp(NumericVector& x, NumericVector& event, NumericMatrix post,
-  String distribution = "weibull", int k = 2, String method = "EM",  int n_iter = 100,
-  double conv_limit = 1e-6) {
-
+                    String distribution = "weibull", int k = 2,
+                    String method = "EM", int n_iter = 100,
+                    double conv_limit = 1e-6) {
 
   int n = x.length();
 
   NumericMatrix posterior = clone(post);
   NumericMatrix logL(n, k); // Likelihood P_ij of the data i belonging to sub-population j
 
-  NumericVector prior(k, 1.0);   // Mixing weight (relevance of cluster)
+  NumericVector prior(k, 1.0); // Mixing weight (relevance of cluster)
   prior = prior / k;
 
   NumericMatrix parameter(2, k);
 
-  // double logLikelihood_old = -std::numeric_limits<double>::infinity();
+  //double logLikelihood_old = -std::numeric_limits<double>::infinity();
   double logLikelihood_old = 0;
-
   for (int iter = 0; iter < n_iter; ++iter) {
 
     //######## M-Step ###########
-    if(method == "EM") {
+    if (method == "EM") {
       parameter = MStepWeibull(x, posterior, event);
     }
-
-    Rcout << parameter << std::endl;
-    Rcout << posterior << std::endl;
-    Rcout << prior << std::endl;
-    Rcout << logLikelihood_old << std::endl;
-
 
     //######## E-Step ###########
     arma::mat posterior_old = as<arma::mat>(posterior);
@@ -233,7 +228,6 @@ List mixture_em_cpp(NumericVector& x, NumericVector& event, NumericMatrix post,
     LikelihoodWeibull(x, parameter, event, prior, posterior, logL);
 
     normalize(posterior);
-
     prior = colMeans(posterior);
 
     arma::mat logDensity = as<arma::mat>(logL);
@@ -241,24 +235,25 @@ List mixture_em_cpp(NumericVector& x, NumericVector& event, NumericMatrix post,
     NumericVector logPrior = log(prior);
     arma::vec logPrio = as<arma::vec>(logPrior);
 
-    double logLikelihood_new = logLikelihood(posterior_old, logDensity, logPrio);
+    double logLikelihood_new =logLikelihood(posterior_old, logDensity, logPrio);
 
-    double convCrit = abs(logLikelihood_new - logLikelihood_old) / (abs(logLikelihood_old) + .001 * conv_limit);
+    double convCrit = (std::abs(logLikelihood_new - logLikelihood_old) /
+                      (std::abs(logLikelihood_old) + 0.001 * conv_limit));
 
     if (convCrit < conv_limit) {
-      Rcout << "For given convergence limit, the algorithm is converged."<< std::endl;
+      // Rcout << "For given convergence limit, the algorithm is converged."<< std::endl;
       logLikelihood_old = logLikelihood_new;
       break;
     }
 
     if (iter == n_iter - 1) {
-      Rcout << "Maximum number of iterations was reached."<< std::endl;
+      Rcpp::warning("Maximum number of iterations was reached.");
     }
 
     logLikelihood_old = logLikelihood_new;
-    }
 
- // return parameter;
+  }
+  // return parameter;
   return List::create(
     _["coefficients"]  = parameter,
     _["posteriori"]  = posterior,